reject candidate condition added

+++ controllers/candidateController.js
@@ -1,6 +1,9 @@
 const { db, ApiError } = require('./baseController')
 const { statuses } = require('../data/enums')
-const { ELECTION_IS_NOT_ACTIVE } = require('../data/errors')
+const {
+  ELECTION_IS_NOT_ACTIVE,
+  CANDIDATE_CANNOT_BE_REJECTED_DURING_ELECTION
+} = require('../data/errors')
 
 exports.getCandidates = async (req, res) => {
   const department = req.auth.department
@@ -39,17 +42,23 @@ exports.beCandidate = async (req, res) => {
     candidate.department = department
 
     candidate = await db.candidate.create(candidate)
-    res.send()
   }
+  res.send()
 }
 
 exports.reject = async (req, res) => {
   const id = req.params.id
+  const department = req.auth.department
 
-  await db.candidate.destroy({
-    where: { id: parseInt(id) }
-  })
+  const election = await db.election.findOne({ where: { department } })
 
+  if (election.status === statuses.ACTIVE) {
+    throw new ApiError(CANDIDATE_CANNOT_BE_REJECTED_DURING_ELECTION)
+  } else {
+    await db.candidate.destroy({
+      where: { id: parseInt(id) }
+    })
+  }
   res.send()
 }
